+++ include/can_frame.h
@@ -6,6 +6,7 @@ namespace canchaos {
 
 struct CanFrame {
   uint32_t id = 0;
+  bool extended = false;
   uint8_t length = 0;
   uint8_t data[8] = {0};
   uint32_t timestampMs = 0;